Major changes to handle the outputting class.
DatabaseInterface.h/.cpp:
Overloaded the constructor so that I can do the outputting of the help
  information from this base class.

Makefile:
Added the stuff for the OutputInterface class.

OutputInterface.h/.cpp
Removed the Callback function, this is unnecessary now.  Added the method
  to handle the outputting of the information, and to handle the calls to
  the outputting methods.

TODO:
Added a todo list.

database.cpp:
Major update to use the new class, and the changes to the base class.
  This also now handles the input parameters.  Things seem to be in
  working order here, but more vigorous checks should be implemented.