Added new class to read data files.
DataReader.cpp/h:
This class will read the data files that are provided.  This eliminates
  the need for some kind of complicated interactive interface.  It will
  also check the modification times of the data files and update the
  database as necessary.

DatabaseInterface.cpp/h:
This guy now opens the database for read/write operations.  When the "-n"
  flag is passed to the program this class will also create a new database
  and create the necessary tables in that database.

Makefile:
Added the new class to the object lists.

OutputInterface.cpp/h:
Some changes to the OutputInformation function so that it can handle the
  single queries and the range queries.  This section still needs to be
  rewritten some, to make the output more tractable.

database.cpp:
Added the new class, and did some better handling of the command line
  arguments.