Major rewrites
Makefile:
Change orders of things.

DatabaseInput.h/.cpp:
The input database now handles creation of the databases.  It no longer
  accesses the database directly at all.  It calls functions in the
  Interface.  Things have been rewritten to handles the new stuff.

DatabaseInterface.h/.cpp:
It does all of the talking to the database.  It will also return the
  results of queries.

DatabaseOutput.h/.cpp:
Now handles help outputting, and has been rewritten to take care of the
  new stuff.

database.cpp:
Uses the getopt function to handle the flags now.  This is a much more
  elegant solution than what was previously done.